Separate TestStatsClient
